services/agent/requirements.txt
# Servicio HTTP
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests>=2.31.0
pydantic>=2.7,<3
pydantic-settings>=2.2,<3

# LangChain + LangGraph
langchain-core>=0.3,<2
langchain-community>=0.3,<0.5
langchain-text-splitters>=0.3,<2
langgraph>=0.2.35,<0.3

# Google Gemini (wrapper LangChain)
langchain-google-genai>=2.1,<3
google-genai>=0.3

# Embeddings / utilidades
langchain-huggingface>=0.1,<2
sentence-transformers>=2.7
pypdf>=4.2

# Web search (Tavily)
tavily-python>=0.4
langchain-tavily>=0.1

# Haystack 2.x + OpenSearch
haystack-ai>=2.0
opensearch-haystack>=4.0
opensearch-py>=2.6
transformers>=4.38
torch>=2.2.0

# Tests del agente
pytest>=8